+++ test/test.py
@@ -129,13 +129,19 @@ def test_ohlcv(exchange, symbol):
 # ------------------------------------------------------------------------------
 
 
-def test_tickers(exchange):
+def test_tickers(exchange, symbol):
     if exchange.hasFetchTickers:
         delay = int(exchange.rateLimit / 1000)
         time.sleep(delay)
-        dump(green(exchange.id), 'fetching all tickers at once...')
-        tickers = exchange.fetch_tickers()
-        dump(green(exchange.id), 'fetched', green(len(list(tickers.keys()))), 'tickers')
+        tickers = None
+        try:
+            dump(green(exchange.id), 'fetching all tickers at once...')
+            tickers = exchange.fetch_tickers()
+            dump(green(exchange.id), 'fetched all', green(len(list(tickers.keys()))), 'tickers')
+        except Exception as e:
+            dump(green(exchange.id), 'failed to fetch all tickers, fetching multiple tickers at once...')
+            tickers = exchange.fetch_tickers([symbol])
+            dump(green(exchange.id), 'fetched', green(len(list(tickers.keys()))), 'tickers')
     else:
         dump(yellow(exchange.id), 'fetching all tickers at once not supported')
 
@@ -187,7 +193,7 @@ def test_symbol(exchange, symbol):
         test_order_book(exchange, symbol)
         test_trades(exchange, symbol)
 
-    test_tickers(exchange)
+    test_tickers(exchange, symbol)
     test_ohlcv(exchange, symbol)
 
 # ------------------------------------------------------------------------------