+++ test/test_async.py
@@ -130,13 +130,19 @@ async def test_ohlcv(exchange, symbol):
 # ------------------------------------------------------------------------------
 
 
-async def test_tickers(exchange):
+def test_tickers(exchange, symbol):
     if exchange.hasFetchTickers:
         delay = int(exchange.rateLimit / 1000)
         time.sleep(delay)
-        dump(green(exchange.id), 'fetching all tickers at once...')
-        tickers = await exchange.fetch_tickers()
-        dump(green(exchange.id), 'fetched', green(len(list(tickers.keys()))), 'tickers')
+        tickers = None
+        try:
+            dump(green(exchange.id), 'fetching all tickers at once...')
+            tickers = exchange.fetch_tickers()
+            dump(green(exchange.id), 'fetched all', green(len(list(tickers.keys()))), 'tickers')
+        except Exception as e:
+            dump(green(exchange.id), 'failed to fetch all tickers, fetching multiple tickers at once...')
+            tickers = exchange.fetch_tickers([symbol])
+            dump(green(exchange.id), 'fetched', green(len(list(tickers.keys()))), 'tickers')
     else:
         dump(yellow(exchange.id), 'fetching all tickers at once not supported')
 
@@ -188,7 +194,7 @@ async def test_symbol(exchange, symbol):
         await test_order_book(exchange, symbol)
         await test_trades(exchange, symbol)
 
-    await test_tickers(exchange)
+    await test_tickers(exchange, symbol)
     await test_ohlcv(exchange, symbol)
 
 # ------------------------------------------------------------------------------