+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Сhangelog
 
+- Yobit and Liqui unified (same BTC-e API)
 - Added support for a new exchange: [Allcoin](https://allcoin.com) (Canada), [#317](https://github.com/kroitor/ccxt/issues/317)
 - Fixed HitBTC v1 market buy/sell, [#314](https://github.com/kroitor/ccxt/issues/314)
 - Fixed PHP 7 method declaraions, [#313](https://github.com/kroitor/ccxt/issues/313)